pyproject.toml
[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"

[project]
name = "telegram-adapter"
version = "0.1.0"
description = "Interface between request from Telegram and the AI Agent"
authors = [{ name = "nicholas" , email = "nicholasvachon@hotmail.com"}]
requires-python = ">=3.13"
dependencies = [
    "langchain-community>=0.3.27",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-telegram-bot>=22.3",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]


[tool.hatch.build.targets.wheel]
    packages = ["src/telegram_adapter"]

